User locations now update in firebase, added methods to pushnotifications and database
